requirements.txt
neo4j==5.28.1
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.0.0
pydantic==2.11.4